Als 4496 update swagger in gitbook to use nhanes (#9)
* Update httpd-vhosts.conf

Added Location element to allow users to load our openapi documentation on a different hosts. This resolves cors issues we would otherwise experience.

* Revert Dockerfile alpine version

This resolves an issue experienced when trying to use the latest version of baseline in nhanes.

* Reverted Dockerfile to alpine:2.4.53

Reverted the Dockerfile alpine version because we have downgraded httpd-vhosts.conf SSLProtocol and SSLProxyProtocol.